Added decrease() and autoDecrease()
The functions _autoDecrease and _decreasePercent were added, as well as
'bound' to progressJs.decrease() and progressJs.autoDecrease().

These functions can, for example, be used to decrease the progressbar
when going back in a slideshow, where the progressbar indicated progress
of presentation.

(I apologise for automatic trimming of whitespace, and hope it doesn't
pose any trouble. )